Use targets selector
This uses the native homeassistant target selector over the previous
entities selector. This means we can target areas as well as individual
lights, and aligns with the expectations of the platform in general.

This commit also includes a refactor of the card configuration that will
migrate previous configurations to the latest state (specifically from
the use of entities to targets for previous version migration)